+++ .github/workflows/main.yml
@@ -31,4 +31,6 @@ jobs:
 
       # Runs a set of commands using the runners shell
       - name: Test
-        run: yarn test
+        run: |
+          yarn lint
+          yarn test
