+++ package.json
@@ -5,10 +5,11 @@
   "repository": {
     "url": "https://github.com/erikmueller/cond-flow"
   },
-  "main": "src/index.js",
+  "main": "built/index.js",
   "scripts": {
     "lint": "eslint src",
-    "test": "yarn lint && jest --collect-coverage"
+    "test": "jest --collect-coverage",
+    "build": "tsc"
   },
   "keywords": [
     "cond",
@@ -26,6 +27,7 @@
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-promise": "5.1.0",
     "jest": "27.2.5",
-    "prettier": "2.4.1"
+    "prettier": "2.4.1",
+    "typescript": "^4.4.4"
   }
 }
